# By urban-warrior (13) and dlemstra (1)
# Via urban-warrior
* 'master' of ssh://github.com/ImageMagick/ImageMagick:
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18852 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18850 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18848 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18847 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18846 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  Fix path to lib dir.
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18841 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18840 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18837 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18835 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18833 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18831 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18828 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74
  git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18827 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/tiff.c
@@ -991,7 +991,7 @@ static void TIFFReadPhotoshopLayers(Image* image,const ImageInfo *image_info,
   info.version=1;
   info.columns=layers->columns;
   info.rows=layers->rows;
-  info.channels=layers->number_channels;
+  info.channels=(unsigned short) layers->number_channels;
   ReadPSDLayers(layers,image_info,&info,MagickFalse,exception);
   DeleteImageFromList(&layers);
   if (layers != (Image *) NULL)
